refactor(finance): consolidate cashAmount and coinAmount into Amount and replace monetary types with Long

+++ src/main/java/com/roy/morago/dto/finance/WalletDTO.java
@@ -1,18 +1,15 @@
 package com.roy.morago.dto.finance;
 
-import com.roy.morago.enums.CurrencyCode;
 import com.roy.morago.enums.WalletStatus;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
+import com.roy.morago.enums.CurrencyCode;
+import lombok.*;
 
 @Getter
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
 public class WalletDTO {
-    private Integer balance;
+    private Long balance;
     private CurrencyCode currencyCode;
     private WalletStatus status;
 }
